+++ src/main/java/SimpleRecommendationSystem/recommendation/api/RecipeRestController.java
@@ -1,7 +1,7 @@
 package SimpleRecommendationSystem.recommendation.api;
 
 import SimpleRecommendationSystem.recommendation.model.Recipe;
-import SimpleRecommendationSystem.recommendation.services.RecipeService;
+import SimpleRecommendationSystem.recommendation.services.RecipeServiceImpl;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -13,15 +13,15 @@ import java.util.List;
 @RequestMapping("/api/v1/recipe")
 public class RecipeRestController {
 
-    private final RecipeService recipeService;
+    private final RecipeServiceImpl recipeServiceImpl;
 
-    public RecipeRestController(RecipeService recipeService) {
-        this.recipeService = recipeService;
+    public RecipeRestController(RecipeServiceImpl recipeServiceImpl) {
+        this.recipeServiceImpl = recipeServiceImpl;
     }
 
     @GetMapping
     public List<Recipe> getRecipe(@RequestParam(required = false) String name, @RequestParam(required = false) String ingredient, @RequestParam(required = false) String instruction) {
-        return recipeService.getRecipes(name, ingredient, instruction);
+        return recipeServiceImpl.getRecipes(name, ingredient, instruction);
     }
 
 }
