+++ src/main/java/SimpleRecommendationSystem/recommendation/repository/IngredientMapper.java
@@ -1,6 +1,7 @@
-package SimpleRecommendationSystem.recommendation.services;
+package SimpleRecommendationSystem.recommendation.repository;
 
 
+import SimpleRecommendationSystem.recommendation.model.Ingredient;
 import org.springframework.jdbc.core.RowMapper;
 
 import java.sql.ResultSet;
