+++ src/main/java/SimpleRecommendationSystem/recommendation/repository/InstructionMapper.java
@@ -1,5 +1,6 @@
-package SimpleRecommendationSystem.recommendation.services;
+package SimpleRecommendationSystem.recommendation.repository;
 
+import SimpleRecommendationSystem.recommendation.model.Instruction;
 import org.springframework.jdbc.core.RowMapper;
 
 import java.sql.ResultSet;
