+++ src/main/java/SimpleRecommendationSystem/recommendation/repository/RecipeMapper.java
@@ -1,6 +1,8 @@
-package SimpleRecommendationSystem.recommendation.services;
+package SimpleRecommendationSystem.recommendation.repository;
 
 import SimpleRecommendationSystem.recommendation.model.Recipe;
+import SimpleRecommendationSystem.recommendation.model.Ingredient;
+import SimpleRecommendationSystem.recommendation.model.Instruction;
 import org.springframework.jdbc.core.RowMapper;
 
 import java.sql.ResultSet;
