+++ src/main/java/SimpleRecommendationSystem/recommendation/repository/RecipeRepository.java
@@ -1,9 +1,8 @@
-package SimpleRecommendationSystem.recommendation.services;
+package SimpleRecommendationSystem.recommendation.repository;
 
 import SimpleRecommendationSystem.recommendation.model.Recipe;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
