--- src/main/java/SimpleRecommendationSystem/recommendation/services/Ingredient.java
@@ -1,19 +1,0 @@
-package SimpleRecommendationSystem.recommendation.services;
-
-public class Ingredient {
-    private final int id;
-    private final String name;
-
-    public Ingredient(int id, String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
