--- src/main/java/SimpleRecommendationSystem/recommendation/services/Instruction.java
@@ -1,19 +1,0 @@
-package SimpleRecommendationSystem.recommendation.services;
-
-public class Instruction {
-    private final int id;
-    private final String name;
-
-    public Instruction(int id, String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
